modified datatypes of number1 and number 2

+++ functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
 
 int	main(void)
 {
-	double	number1 = 1, number2 = 2, number3;
+	unsigned int	number1 = 1, number2 = 2, number3;
 	int	index = 3;
 
-	printf("%.0f, %.0f", number1, number2);
+	printf("%u, %u", number1, number2);
 	for (; index <= 98; index++)
 	{
 		number3 = number1 + number2;
-		printf(", %.0f", number3);
+		printf(", %u", number3);
 		number1 = number2;
 		number2 = number3;
 	}
